apps/ai-service/requirements.txt
# Core FastAPI framework
fastapi==0.109.2
uvicorn==0.27.1

# Data validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
pandas==2.2.0

# OCR and image processing
pytesseract==0.3.10
Pillow==10.0.1

# AI and embeddings
sentence-transformers==2.5.1
numpy==1.26.4

# Vector database
qdrant-client==1.14.3

# Cryptography for token decryption
cryptography!=3.4.0

# HTTP requests for logging
requests

# Cloud deployment (for modal_service_persistent.py)
modal

# Optional: For development and debugging
python-multipart==0.0.9

# Note: SQLAlchemy and database dependencies removed since AI service doesn't manage database records